feat: supported gpt4o models (#107)
