+++ .ort.yml
@@ -19,9 +19,9 @@ resolutions:
     - message: ".*PyPI::numpy:1\\.26\\.0.*"
       reason: "CANT_FIX_EXCEPTION"
       comment: "BSD 3-Clause license: https://github.com/numpy/numpy/blob/v1.26.0/LICENSES_bundled.txt"
-    - message: ".*PyPI::tiktoken:0\\.5\\.1.*"
+    - message: ".*PyPI::tiktoken:0\\.7\\.0.*"
       reason: "CANT_FIX_EXCEPTION"
-      comment: "MIT License: https://github.com/openai/tiktoken/blob/0.5.1/LICENSE"
+      comment: "MIT License: https://github.com/openai/tiktoken/blob/0.7.0/LICENSE"
     - message: ".*PyPI::httpcore:0\\.18\\.0.*"
       reason: "CANT_FIX_EXCEPTION"
       comment: "BSD 3-Clause New or Revised License: https://github.com/encode/httpcore/blob/0.18.0/LICENSE.md"
