+++ aidial_adapter_openai/gpt4_multi_modal/messages.py
